Renames components extension to jsx
