+++ client/public/styles/index.scss
@@ -47,7 +47,7 @@ li.active {
     margin-top: 15px
 }
 .loginForm {
-    margin-top: 100px
+    margin-top: 50px
 }
 .confirmDelete, .replyReview {
     margin-right: 15px
