+++ client/src/components/Home/presentational/HomeImageSlider.jsx
