+++ client/src/components/Home/presentational/HowItWorks.jsx
