+++ client/src/components/NavBar/presentational/GuestNavBar.jsx
