+++ client/src/containers/businessProfile.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import BusinessProfile from '../components/Businesses/smart/BusinessProfile';
+import BusinessProfile from '../components/Businesses/smart/BusinessProfile.jsx';
 import { fetchBusiness } from '../store/actions/businesses';
 import { logout } from '../store/actions/auth';
 import { fetchReviews, postReview, postReviewResponse } from '../store/actions/businessReview';
