+++ server/controllers/Businesses.js
@@ -39,7 +39,6 @@ export default class Businesses {
       location: req.body.location,
       address: req.body.address,
       userId: req.userData.userId,
-      businessOwner: req.userData.username,
       description: req.body.description
     };
     return Business
@@ -55,7 +54,6 @@ export default class Businesses {
           homeNumber: business.homeNumber,
           location: business.location,
           address: business.address,
-          businessAddedBy: business.businessOwner,
           description: business.description
         };
         res.status(201)
@@ -121,7 +119,7 @@ export default class Businesses {
         include: [
           {
             model: User,
-            as: 'user',
+            as: 'businessOwner',
             attributes: ['username']
           }
         ]
@@ -176,7 +174,12 @@ export default class Businesses {
       .findOne({
         where: {
           id: req.params.businessId
-        }
+        },
+        include: [{
+          model: User,
+          as: 'businessOwner',
+          attributes: ['username']
+        }]
       })
       .then((business) => {
         if (!business) {
@@ -206,7 +209,7 @@ export default class Businesses {
               homeNumber: updatedBusiness.homeNumber,
               location: updatedBusiness.location,
               address: updatedBusiness.address,
-              businessAddedBy: updatedBusiness.user.username,
+              businessAddedBy: updatedBusiness.businessOwner.username,
               description: updatedBusiness.description
             };
             res.status(200).json({ message: 'Business Updated successfully', updatedBusinessDetails });
@@ -264,7 +267,6 @@ export default class Businesses {
     const businessReviewDetails = {
       reviewerId: req.userData.userId,
       review: req.body.review,
-      reviewer: req.userData.username,
       businessId: req.params.businessId
     };
     return Business
@@ -313,7 +315,6 @@ export default class Businesses {
     const businessReviewResponse = {
       userId: req.userData.userId,
       message: req.body.message,
-      reviewer: req.userData.username,
       reviewId: req.params.reviewId
     };
     return BusinessReview
@@ -368,12 +369,13 @@ export default class Businesses {
                 as: 'responses',
                 include: [{
                   model: User,
-                  as: 'user'
+                  as: 'reviewer',
+                  attributes: ['username']
                 }]
               },
               {
                 model: User,
-                as: 'user',
+                as: 'reviewer',
                 attributes: ['username']
               }
               ],
