+++ server/models/business.js
@@ -28,10 +28,6 @@ const business = (sequelize, DataTypes) => {
         // }
       }
     },
-    businessOwner: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
     businessImage: {
       type: DataTypes.STRING,
       allowNull: true,
@@ -188,7 +184,7 @@ const business = (sequelize, DataTypes) => {
     Business.belongsTo(models.User, {
       foreignKey: 'userId',
       onDelete: 'CASCADE',
-      as: 'user'
+      as: 'businessOwner'
     });
   };
   return Business;
