+++ server/models/reviewresponse.js
@@ -6,10 +6,6 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.UUID,
       defaultValue: DataTypes.UUIDV4,
     },
-    reviewer: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
     message: {
       type: DataTypes.STRING,
       allowNull: false,
@@ -33,7 +29,7 @@ module.exports = (sequelize, DataTypes) => {
     reviewresponse.belongsTo(models.User, {
       foreignKey: 'userId',
       onDelete: 'CASCADE',
-      as: 'user'
+      as: 'reviewer'
     });
   };
   return reviewresponse;
